simplified search stream creation
